+++ src/main/java/com/braintreegateway/PayPalDetails.java
@@ -17,6 +17,9 @@ public class PayPalDetails {
     private String sellerProtectionStatus;
     private String captureId;
     private String refundId;
+    private String transactionFeeAmount;
+    private String transactionFeeCurrencyIsoCode;
+    private String description;
 
     public PayPalDetails(NodeWrapper node) {
         payerEmail = node.findString("payer-email");
@@ -33,6 +36,9 @@ public class PayPalDetails {
         sellerProtectionStatus = node.findString("seller-protection-status");
         refundId = node.findString("refund-id");
         captureId = node.findString("capture-id");
+        transactionFeeAmount = node.findString("transaction-fee-amount");
+        transactionFeeCurrencyIsoCode = node.findString("transaction-fee-currency-iso-code");
+        description = node.findString("description");
     }
 
     public String getPayerEmail() {
@@ -89,5 +95,17 @@ public class PayPalDetails {
 
     public String getRefundId() {
       return refundId;
+    }
+
+    public String getTransactionFeeAmount() {
+      return transactionFeeAmount;
+    }
+
+    public String getTransactionFeeCurrencyIsoCode() {
+      return transactionFeeCurrencyIsoCode;
+    }
+
+    public String getDescription() {
+      return description;
     }
 }