+++ src/test/java/com/braintreegateway/integrationtest/OAuthIT.java
@@ -216,4 +216,13 @@ public class OAuthIT {
             fail("malformed url");
         }
     }
+
+    @Test
+    public void computeSignatureReturnsCorrectSignature() {
+        String url = "http://localhost:3000/oauth/connect?business%5Bname%5D=We+Like+Spaces&client_id=client_id%24development%24integration_client_id";
+
+        String signature = gateway.oauth().computeSignature(url);
+
+        assertEquals("a36bcf10dd982e2e47e0d6a2cb930aea47ade73f954b7d59c58dae6167894d41", signature);
+    }
 }