+++ src/test/java/com/braintreegateway/integrationtest/TransactionIT.java
@@ -3583,6 +3583,27 @@ public class TransactionIT implements MerchantAccountTestConstants {
     }
 
     @Test
+    public void createPayPalTransactionWithPayPalDescription() {
+        String nonce = TestHelper.generateOneTimePayPalNonce(gateway);
+        TransactionRequest request = new TransactionRequest().
+            amount(new BigDecimal("100.00")).
+            paymentMethodNonce(nonce).
+            paypalAccount().
+              done().
+            options().
+              paypal().
+                description("Product description").
+                done().
+              done();
+
+        Result<Transaction> saleResult = gateway.transaction().sale(request);
+
+        assertTrue(saleResult.isSuccess());
+        assertNotNull(saleResult.getTarget().getPayPalDetails());
+        assertEquals("Product description", saleResult.getTarget().getPayPalDetails().getDescription());
+    }
+
+    @Test
     public void createOneTimePayPalTransactionAndAttemptToVault() {
         String nonce = TestHelper.generateOneTimePayPalNonce(gateway);
         TransactionRequest request = new TransactionRequest().
@@ -3735,6 +3756,8 @@ public class TransactionIT implements MerchantAccountTestConstants {
         assertNotNull(transaction.getPayPalDetails().getSellerProtectionStatus());
         assertNotNull(transaction.getPayPalDetails().getCaptureId());
         assertNotNull(transaction.getPayPalDetails().getRefundId());
+        assertNotNull(transaction.getPayPalDetails().getTransactionFeeAmount());
+        assertNotNull(transaction.getPayPalDetails().getTransactionFeeCurrencyIsoCode());
     }
 
     @Test